refactor: split ItemDetailView into separate presentation file
- Created ItemDetailViewPresentation.tsx with pure presentation logic
- No Meteor or Wouter dependencies in presentation component
- ItemDetailView.tsx now only contains route-aware wrapper
- Re-exports presentation component for backward compatibility
- Updated ItemDetailView.stories.tsx to import from presentation file
- Fixes Storybook webpack build errors (can't resolve meteor/meteor)

This follows the same pattern used for AllItemsView and AllTagsView.
Storybook build now completes successfully without webpack errors.
