+++ specs/003-url-routing/plan.md
@@ -247,9 +247,8 @@ const { itemId } = useParams<{ itemId: string }>();
 - These are test maintenance issues unrelated to routing functionality
 - **Routing implementation is complete and working**
 
-**Conclusion**: 
+**Conclusion**:
 - ✅ URL routing successfully implemented
-- ✅ `/tags` route navigates and renders correctly  
+- ✅ `/tags` route navigates and renders correctly
 - ✅ T012 from spec-002 is now **UNBLOCKED**
 - ⚠️ Test selectors need updating to match actual UI (separate maintenance task)
-
